total shit

+++ app/src/main/java/cz/upce/bdats/autopujcovna/OsobniAuto.java
@@ -6,10 +6,10 @@ public class OsobniAuto extends Auto {
 
     // Konstruktor
     public OsobniAuto(String spz, Barva barva) {
-        this(spz, 0.0f, 0, barva);
+        this(spz, 0, 0, barva);
     }
 
-    public OsobniAuto(String spz, float stavKm, int pocetVypujceni, Barva barva) {
+    public OsobniAuto(String spz, int stavKm, int pocetVypujceni, Barva barva) {
         super(Typ.OSOBNI, spz, stavKm, pocetVypujceni);
         this.barva = barva;
     }
