+++ app/src/main/java/cz/upce/bdats/autopujcovna/UzitkoveAuto.java
@@ -6,10 +6,10 @@ public class UzitkoveAuto extends Auto {
 
     // Kontruktor
     public UzitkoveAuto(String spz, float nosnost) {
-        this(spz, 0.0f, 0, nosnost);
+        this(spz, 0, 0, nosnost);
     }
 
-    public UzitkoveAuto(String spz, float stavKm, int pocetVypujceni, float nosnost) {
+    public UzitkoveAuto(String spz, int stavKm, int pocetVypujceni, float nosnost) {
         super(Typ.UZITKOVY, spz, stavKm, pocetVypujceni);
         this.nosnost = nosnost;
     }
